refactor(mint): rework Balance split variants to minimal, low-density layouts
Reframe the three variants around subtraction, not restyling. The previous
variants (and the original screen) were too cluttered; every variant is now
visibly calmer and lower-density, with no expand/collapse (which would add
state and hide the part-to-whole relationship).

Variants (enum renamed list | total | donut, default list):
- list:  a bare slider list — nothing wrapping the rows
- total: one calm "to split" figure over the same list
- donut: a proportion donut (shape only) over the same list

All three are built from one shared MintShareRow (mint icon + muted name +
loud % + slider). Subtraction performed (vs the original):
- removed: help-icon Alert (~40 words), "Total distribution" label, info Card
  (~30 words), per-mint card + 4 LinearGradient accent overlays, per-mint
  balance line, Min/Max buttons, the avatar proportion bar, % decimals
- removed actions: Reset (mirror) and Focus (concentrate) — store actions
  remain but are no longer surfaced here; Equalize kept as "Split evenly"
- kept: currency tabs, mint icon + name, slider, %, a single ambient validity
  "100%" (coloured only when off), Split evenly, Next
- net: ~11 per-mint elements -> 4; screen copy ~120 words -> ~2

Deleted the now-unused MintDistributionItem-era components (DistributionBar,
MintSliderControls) and the old variant files; trimmed useDistributionRow to
the slider seam; the % uses the canonical bpToPercent helper.
